refactor: rename fs executor mods
To make it more clear.
