+++ src/connector/src/with_options.rs
@@ -22,7 +22,7 @@ use crate::source::cdc::MYSQL_CDC_CONNECTOR;
 use crate::source::iceberg::ICEBERG_CONNECTOR;
 use crate::source::{
     AZBLOB_CONNECTOR, GCS_CONNECTOR, KAFKA_CONNECTOR, OPENDAL_S3_CONNECTOR, POSIX_FS_CONNECTOR,
-    UPSTREAM_SOURCE_KEY,
+    S3_CONNECTOR, UPSTREAM_SOURCE_KEY,
 };
 
 /// Marker trait for `WITH` options. Only for `#[derive(WithOptions)]`, should not be used manually.
@@ -151,6 +151,12 @@ pub trait WithPropertiesExt: Get + Sized {
     fn connector_need_pk(&self) -> bool {
         // Currently only iceberg connector doesn't need primary key
         !self.is_iceberg_connector()
+    }
+
+    fn is_legacy_fs_connector(&self) -> bool {
+        self.get(UPSTREAM_SOURCE_KEY)
+            .map(|s| s.eq_ignore_ascii_case(S3_CONNECTOR))
+            .unwrap_or(false)
     }
 
     fn is_new_fs_connector(&self) -> bool {
